Allow user configurable socket path
The default socket path (/tmp/manhole-pid) may be inconvient in
environment that do not manage pid files, for example when using systemd
services. Also, using world writable directory (/tmp) for the socket may
be racy due to symlink attacks.

This patch adds an option for user configuratble socket path. This
allows an application to create the socket in a safer and more convient
location such as /var/run/appname/manhole.