Extract connect_to_manhole() helper function
Testing manhole connection was performed by creating a new connection to
manhole. This works for the current code, but for the next patch, I
need a way to create a connection to manhole, keep it open, and check
that the connection is good after fork.

This patch extract a connect_to_manhole() helper function, which is used
now only by assert_manhole_is_running(). This function is used in the
next patch to keep a connection to manhole during the test.

This also fix few minor issues:

- sock was not closed after connection timeout
- sock.shutdown() and sock.close() were not needed in
  assert_manhole_is_running(), since TestSocket() shutdown and close the
  socket when exiting from the scope
- if connection timed out, there is no need to sleep before raising