Do not patch fork if using socket_path option
The socket_path option is not compatible with automatic manhole
installation after fork. The socket of the parent is removed and
replaced with the socket of the child.  Users of socket_path should use
patch_fork=False. If manhole for child process is needed, it must be
installed manually using a different socket_path.

This patch disables fork patching when both patch_fork and socket_path
are used, and print a warning message. It would be easier to debug if an
exception is raised in such case, but this is how other options
conflicts are handled.