Fix leftover /tmp/manhole-pid sockets in the tests
manhole.install() registers an exit function removing the manhole
socket, but atexit function do not run when process is terminated by an
unhandled signal.

In the fork tests, we avoid this issue by terminating child processes
with SIGINT, however the test child processes, created by the
process_tests library, are terminate using SIGTERM in
TestProcess.__exit__.

Adding SIGTERM signal handler fixed this issue.

We probably need to document this as a workaround for issue #8