Make ALL_SIGNALS immutable
ALL_SIGNALS is passed to install() as default value, and may be modified
later by the called functions. This may change the value of the gloabl,
which will change the behavior on the next install() call. Currnetly
this code is safe becuase intall() can be called only once, but once we
introduce uninstall(), this code is not safe anymore.

Now we use a tuple to make sure that the gloabl cannot be changed, and
bad code that try to change will fail loudly.