Remove unneeded poll flags
According to poll(2) POLLERR and POLLHUP are are meaningless in the
events field, and will be set in the revents field whenever the
corresponding condition is true. So there is no need to register for
these events.

POLLPRI means there is urgent data to read. This can be read using the
MSG_OOB flag, but manhole-cli does not use this feature, so there is no
need to wait for this event.